# crates/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Shelfy API",
    "description": "HTTP API of Shelfy Web. JSON in camelCase, timestamps in unix milliseconds. Errors are `application/problem+json` documents whose `code` is stable; clients map it to their own messages.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "1"
  },
  "paths": {
    "/api/v1/auth/device/approve": {
      "post": {
        "tags": ["auth"],
        "summary": "Approves a device sign-in for the signed-in account: the device that\nshows this code gets a `migrate` token for this account's library, valid\n7 days.",
        "description": "Needs a sign-in or a re-authentication from the last 5 minutes (403\n`reauth_required` otherwise). 400 `invalid_device_code` for a code that\nis unknown, expired, used or approved by another account; approving a\ncode again is a no-op. Limit: 10 tries per 10 minutes per account (429).",
        "operationId": "approveDeviceSignIn",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceApproval"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Approved: the device's next poll gets its token."
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/api/v1/auth/device/poll": {
      "post": {
        "tags": ["auth"],
        "summary": "Asks whether a device sign-in was approved.",
        "description": "`pending`: wait `interval` seconds and poll again. `slow_down`: the poll\ncame too soon; wait the new `interval`. `approved`: the `migrate` token,\ndelivered once; the codes stop working. 400 `invalid_device_code` for a\ndevice code that is unknown, expired or used: start again. 429\n`rate_limited` with `Retry-After` past 20 polls a minute of one device\ncode. The sign-in limit per client does not count polls. No cookie and no\nCSRF headers are needed.",
        "operationId": "pollDeviceSignIn",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevicePollRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Where the sign-in stands.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevicePoll"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "security": [{}]
      }
    },
    "/api/v1/auth/device/start": {
      "post": {
        "tags": ["auth"],
        "summary": "Starts signing a device in: the migration CLI's `login`.",
        "description": "Show the user `userCode` and `verificationUri` (or open\n`verificationUriComplete`), then poll `POST /auth/device/poll` with\n`deviceCode` every `interval` seconds until it is approved, for at most\n`expiresIn` seconds. Limit: the sign-in limit, 10 requests per minute per\nclient over every `/auth` route (429). No cookie and no CSRF headers are\nneeded.",
        "operationId": "startDeviceSignIn",
        "responses": {
          "200": {
            "description": "The codes of a new device sign-in.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceAuthorization"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "security": [{}]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": ["auth"],
        "summary": "Signs out: ends the session of the request's cookie and clears the\ncookie. Answers 204 even without a session; without a session cookie the\nanswer sets no cookie.",
        "operationId": "logout",
        "responses": {
          "204": {
            "description": "Signed out; the session cookie, if sent, is cleared."
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "security": [{}]
      }
    },
...
    "/api/v1/auth/magic-links": {
      "post": {
        "tags": ["auth"],
        "summary": "Emails a sign-in link to the address, if it belongs to an account.",
        "description": "The answer is 202 whether or not the account exists, and whether or not\nemail is configured (see `GET /auth/methods`). The email carries the link\n`<public url>/login/magic#<token>`. Limits: 10 requests per minute per\nclient and 3 per hour per address (429 `rate_limited` with\n`Retry-After`). A malformed address answers 422 `validation_failed`.",
        "operationId": "requestMagicLink",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MagicLinkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted. If the address has an account, a link valid for 15 minutes is on its way."
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "security": [{}]
      }
    },
    "/api/v1/auth/magic-links/redeem": {
      "post": {
        "tags": ["auth"],
        "summary": "Redeems a sign-in link: the SPA page `/login/magic#<token>` sends the\ntoken from its URL's fragment, which never reaches a server log.",
        "description": "A usable link starts a session (replacing the one the browser held);\notherwise 400 `invalid_link`. The link works once. Limit: 10 requests per\nminute per client, shared with `POST /auth/magic-links`.",
        "operationId": "redeemMagicLink",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedeemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Signed in: the response sets the session cookie."
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "security": [{}]
      }
    },
...
    "/api/v1/extension/pair": {
      "post": {
        "tags": ["extension"],
        "summary": "Exchanges a pairing code for the extension's token.",
        "description": "The extension calls it with the code that the web app got from `POST\n/me/tokens/pairing-code`, within 60 seconds, once. No cookie and no CSRF\nheaders are needed, and none is read. 400 `invalid_pairing_code` for a\ncode that is malformed, unknown, used or expired; 422 for a malformed\n`installId`, `label` or `version`; 426 `extension_outdated` for a version\nbelow `minVersion` (the code stays usable); 409 `conflict` when the\naccount holds 50 working tokens. Limit: the sign-in limit, 10 requests a\nminute per client over this route and the `/auth` routes (429).",
        "operationId": "pairExtension",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtensionPairRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Paired: the token, shown once.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairedExtension"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "security": [{}]
      }
    },
...
    "/api/v1/me/tokens": {
      "get": {
        "tags": ["account"],
        "summary": "The account's working API tokens (not revoked, not expired), with their\nlast use.",
        "operationId": "listApiTokens",
        "responses": {
          "200": {
            "description": "The tokens.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiTokenList"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      },
      "post": {
        "tags": ["account"],
        "summary": "Creates an API token for the extension or the iOS Shortcut; the answer\nis the only time its value is shown.",
        "description": "Needs a sign-in or a re-authentication from the last 5 minutes (403\n`reauth_required` otherwise). 422 for the kind `migrate` (the migration\nCLI signs in with the device flow), for scopes that are not the kind's,\nor for a label over 64 characters; 409 `conflict` when the account holds\n50 working tokens already.",
        "operationId": "createApiToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "The new token, with its value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedApiToken"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/api/v1/me/tokens/pairing-code": {
      "post": {
        "tags": ["account"],
        "summary": "Creates a code that pairs the browser extension with this account.",
        "description": "The web app hands the code to the extension (`chrome.runtime.sendMessage`),\nwhich exchanges it for its token at `POST /extension/pair` within 60\nseconds, once. Needs a sign-in or a re-authentication from the last 5\nminutes (403 `reauth_required` otherwise). 429 `rate_limited` while the\naccount holds 10 unused codes that have not expired.",
        "operationId": "createPairingCode",
        "responses": {
          "201": {
            "description": "The code and its expiry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairingCode"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/api/v1/me/tokens/{id}": {
      "delete": {
        "tags": ["account"],
        "summary": "Revokes one of the account's tokens: it stops working at once. Another\naccount's token, or one that no longer works, is a 404.",
        "operationId": "revokeApiToken",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The token's id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Revoked."
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiTokenList": {
        "type": "object",
        "description": "The account's working API tokens.",
        "required": ["items"],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiToken"
            },
            "description": "Newest first."
          }
        }
      },
      "ApiTokenRequest": {
        "type": "object",
        "description": "Body of `POST /api/v1/me/tokens`.",
        "required": ["kind"],
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/TokenKind",
            "description": "Who will hold it: `extension` or `shortcut`."
          },
          "label": {
            "type": "string",
            "description": "A name for the list, up to 64 characters; trimmed, and blank is none."
          },
          "scopes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TokenScope"
            },
            "description": "What it may do: some of the kind's scopes; all of them when left out."
          }
        },
        "additionalProperties": false
      },
...
      "CreatedApiToken": {
        "type": "object",
        "description": "A new API token.",
        "required": ["token", "apiToken"],
        "properties": {
          "apiToken": {
            "$ref": "#/components/schemas/ApiToken",
            "description": "The token as the list shows it."
          },
          "token": {
            "type": "string",
            "description": "The value (`shx_…`), for `Authorization: Bearer`. Shown only now:\nthe server keeps its hash."
          }
        }
      },
      "DeviceApproval": {
        "type": "object",
        "description": "Body of `POST /api/v1/auth/device/approve`.",
        "required": ["userCode"],
        "properties": {
          "userCode": {
            "type": "string",
            "description": "The code the CLI shows (`BCDF-GHJK`)."
          }
        }
      },
      "DeviceAuthorization": {
        "type": "object",
        "description": "What `POST /auth/device/start` answers: RFC 8628's device authorization\nresponse, in camelCase.",
        "required": [
          "deviceCode",
          "userCode",
          "verificationUri",
          "verificationUriComplete",
          "expiresIn",
          "interval"
        ],
        "properties": {
          "deviceCode": {
            "type": "string",
            "description": "The CLI's secret: it polls with it. Never show or log it."
          },
          "expiresIn": {
            "type": "integer",
            "format": "int64",
            "description": "Seconds both codes stay valid.",
            "minimum": 0
          },
          "interval": {
            "type": "integer",
            "format": "int64",
            "description": "Seconds to wait between polls.",
            "minimum": 0
          },
          "userCode": {
            "type": "string",
            "description": "What the user approves: 8 letters as `XXXX-XXXX`; case, dashes and\nspaces do not matter."
          },
          "verificationUri": {
            "type": "string",
            "description": "The page that approves it: `<public url>/device`."
          },
          "verificationUriComplete": {
            "type": "string",
            "description": "The same page with the code filled in: `<public url>/device#XXXX-XXXX`."
          }
        }
      },
      "DevicePoll": {
        "oneOf": [
          {
            "type": "object",
            "description": "Not approved yet: poll again in `interval` seconds.",
            "required": ["interval", "status"],
            "properties": {
              "interval": {
                "type": "integer",
                "format": "int64",
                "description": "Seconds to wait before the next poll.",
                "minimum": 0
              },
              "status": {
                "type": "string",
                "enum": ["pending"]
              }
            }
          },
          {
            "type": "object",
            "description": "Polled sooner than the interval allows: wait `interval` seconds from\nnow on (it grew by 5).",
            "required": ["interval", "status"],
            "properties": {
              "interval": {
                "type": "integer",
                "format": "int64",
                "description": "Seconds to wait before the next poll.",
                "minimum": 0
              },
              "status": {
                "type": "string",
                "enum": ["slow_down"]
              }
            }
          },
          {
            "type": "object",
            "description": "Approved: the token, valid 7 days, delivered once. Keep it in a file\nonly its owner can read.",
            "required": ["token", "tokenId", "scopes", "expiresAt", "status"],
            "properties": {
              "expiresAt": {
                "type": "integer",
                "format": "int64",
                "description": "When it stops working, unix ms."
              },
              "scopes": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/TokenScope"
                },
                "description": "What it may do: `[\"migrate\"]`."
              },
              "status": {
                "type": "string",
                "enum": ["approved"]
              },
              "token": {
                "type": "string",
                "description": "The `migrate` token (`shx_…`), for `Authorization: Bearer`."
              },
              "tokenId": {
                "type": "string",
                "description": "Its id, as the account's token list names it."
              }
            }
          }
        ],
        "description": "What a poll finds, by `status`."
      },
      "DevicePollRequest": {
        "type": "object",
        "description": "Body of `POST /api/v1/auth/device/poll`.",
        "required": ["deviceCode"],
        "properties": {
          "deviceCode": {
            "type": "string",
            "description": "The `deviceCode` of `POST /auth/device/start`."
          }
        }
      },
...
      "ExtensionPairRequest": {
        "type": "object",
        "description": "Body of `POST /api/v1/extension/pair` (contract C2). Fields a later\nextension adds are ignored.",
        "required": ["code", "installId", "version"],
        "properties": {
          "code": {
            "type": "string",
            "description": "The pairing code the web app handed over (43 characters)."
          },
          "installId": {
            "type": "string",
            "description": "A random id the extension keeps while it is installed (16 to 128\nletters, digits, `-` or `_`). Pairing the same installation again\nrevokes the token it held."
          },
          "label": {
            "type": "string",
            "description": "A name for the account's token list, such as \"Chrome on macOS\"; up\nto 64 characters, blank is none."
          },
          "version": {
            "type": "string",
            "description": "The extension's manifest version (`0.2.0`); below `minVersion`, 426."
          }
        }
      },
...
      "FieldError": {
        "type": "object",
        "description": "One field that failed validation.",
        "required": ["field", "reason"],
        "properties": {
          "field": {
            "type": "string",
            "description": "The field, as named in the request (camelCase)."
          },
          "reason": {
            "type": "string",
            "description": "Why the value was refused, for developers."
          }
        }
      },
...
      "MagicLinkRequest": {
        "type": "object",
        "description": "Body of `POST /api/v1/auth/magic-links`.",
        "required": ["email"],
        "properties": {
          "email": {
            "type": "string",
            "description": "The account's email address."
          }
        }
      },
...
      "PairedExtension": {
        "type": "object",
        "description": "A paired extension's token.",
        "required": ["token", "tokenId", "scopes"],
        "properties": {
          "scopes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TokenScope"
            },
            "description": "What it may do: `ingest`, `tasks`, `uploads`, `lookup`."
          },
          "token": {
            "type": "string",
            "description": "The `extension` token (`shx_…`) for `Authorization: Bearer`. Shown\nonly now; keep it where page scripts cannot read it."
          },
          "tokenId": {
            "type": "string",
            "description": "Its id, as the account's token list names it."
          }
        }
      },
      "PairingCode": {
        "type": "object",
        "description": "A code that pairs the browser extension (contract C2).",
        "required": ["code", "expiresAt"],
        "properties": {
          "code": {
            "type": "string",
            "description": "The code (43 characters), for the extension's `POST /extension/pair`.\nHand it to the extension only; it works once."
          },
          "expiresAt": {
            "type": "integer",
            "format": "int64",
            "description": "When it stops working, unix ms: 60 seconds from now."
          }
        }
      },
...
      "Problem": {
        "type": "object",
        "description": "An error body: RFC 9457 problem details plus the stable `code`.",
        "required": ["type", "title", "status", "code"],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode",
            "description": "What went wrong. Clients map it to their own message."
          },
          "detail": {
            "type": "string",
            "description": "Developer-facing detail. Never UI prose, user content or secrets."
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "description": "The fields that failed validation (`validation_failed`)."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "The HTTP status of the response.",
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "The reason phrase of `status`. Not meant for display."
          },
          "type": {
            "type": "string",
            "description": "Always `about:blank`: `code` carries the problem type."
          }
        }
      },
...
      "RedeemRequest": {
        "type": "object",
        "description": "Body of `POST /api/v1/auth/magic-links/redeem`.",
        "required": ["token"],
        "properties": {
          "token": {
            "type": "string",
            "description": "The link's token: what follows `#` in its URL."
          }
        }
      },
...
      "TokenKind": {
        "type": "string",
        "description": "Who holds a token.",
        "enum": ["extension", "shortcut", "migrate"]
      },
      "TokenScope": {
        "type": "string",
        "description": "What a token may do (§2.9).",
        "enum": ["ingest", "tasks", "uploads", "lookup", "links:create", "migrate"]
      },
...
    },
    "responses": {
      "Problem": {
        "description": "An error, as RFC 9457 problem details with a stable `code`.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "shx_",
        "description": "A scoped API token (`shx_…`) of the extension, the iOS Shortcut or the migration CLI. Tokens cannot call routes that only accept the session."
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-shelfy_session",
        "description": "The web app's session, set by signing in. Requests that change state (POST, PUT, PATCH, DELETE) with this cookie must also send `X-Shelfy-Client: web` and an `Origin` equal to the public URL; otherwise they fail with 403 `csrf_failed`. A request with an `Authorization` header is never authenticated by this cookie."
      }
    }
  },
  "security": [
    {
      "session": []
    }
  ],
...
}